fix: standardize package naming in treports-samples
Rename package com.tjf.samples.treports to com.tjf.sample.github.treports
across all three treports sub-projects (provider, server-client, viewer),
aligning with the com.tjf.sample.github.* convention used in all other samples.

- Updated package declarations in 30 Java files
- Moved directory structure from com/tjf/samples/ to com/tjf/sample/github/
- Updated all internal imports accordingly
