+++ treports-samples/tjf-sample-treports-provider/src/main/java/com/tjf/sample/github/treports/TreportsApplication.java
@@ -1,0 +1,16 @@
+package com.tjf.sample.github.treports;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+
+
+@SpringBootApplication
+@EnableJpaRepositories()
+public class TreportsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TreportsApplication.class, args);
+	}
+}
