+++ treports-samples/tjf-sample-treports-provider/src/main/java/com/tjf/sample/github/treports/model/Product.java
@@ -1,0 +1,114 @@
+package com.tjf.sample.github.treports.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@Entity
+public class Product {
+
+	@Id
+	private String id;
+	
+	private boolean active;
+	private String branchId;
+	private String brandId;
+	private String companyId;
+	private String createDate;
+	private String description;
+	private String keywords;
+	private String lastChange;
+	private String name;
+	private String title;
+
+	//@JsonProperty("CompanyId")
+	public String getCompanyId() {
+		return companyId;
+	}
+
+	public void setCompanyId(String companyId) {
+		this.companyId = companyId;
+	}
+
+	//@JsonProperty("BranchId")
+	public String getBranchId() {
+		return branchId;
+	}
+
+	public void setBranchId(String branchId) {
+		this.branchId = branchId;
+	}
+
+	//@JsonProperty("BrandId")
+	public String getBrandId() {
+		return brandId;
+	}
+
+	public void setBrandId(String brandId) {
+		this.brandId = brandId;
+	}
+
+	//@JsonProperty("CreateDate")
+	public String getCreateDate() {
+		return createDate;
+	}
+
+	public void setCreateDate(String createDate) {
+		this.createDate = createDate;
+	}
+
+	//@JsonProperty("LastChange")
+	public String getLastChange() {
+		return lastChange;
+	}
+
+	public void setLastChange(String lastChange) {
+		this.lastChange = lastChange;
+	}
+
+	//@JsonProperty("Description")
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	//@JsonProperty("isActive")
+	public boolean isActive() {
+		return active;
+	}
+
+	public void setActive(boolean active) {
+		this.active = active;
+	}
+
+	//@JsonProperty("Keywords")
+	public String getKeywords() {
+		return keywords;
+	}
+
+	public void setKeywords(String keywords) {
+		this.keywords = keywords;
+	}
+
+	//@JsonProperty("Name")
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	//@JsonProperty("Title")
+	public String getTitle() {
+		return title;
+	}
+
+	public void setTitle(String title) {
+		this.title = title;
+	}
+}
