+++ treports-samples/tjf-sample-treports-provider/src/main/java/com/tjf/sample/github/treports/model/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.tjf.sample.github.treports.model;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.totvs.tjf.api.jpa.repository.ApiJpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, String>, ApiJpaRepository<Product>{
+}
