--- treports-samples/tjf-sample-treports-provider/src/main/java/com/tjf/samples/treports/model/ProductRepository.java
@@ -1,8 +1,0 @@
-package com.tjf.samples.treports.model;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.totvs.tjf.api.jpa.repository.ApiJpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, String>, ApiJpaRepository<Product>{
-}
