+++ treports-samples/tjf-sample-treports-server-client/src/main/java/com/tjf/sample/github/treports/Application.java
@@ -1,0 +1,12 @@
+package com.tjf.sample.github.treports;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+}
