+++ treports-samples/tjf-sample-treports-server-client/src/main/java/com/tjf/sample/github/treports/api/ApiExecuteInfoResponse.java
@@ -1,0 +1,23 @@
+package com.tjf.sample.github.treports.api;
+
+public class ApiExecuteInfoResponse {
+
+	private String uIdRequest;
+	private String scheduleDate;
+
+	public String getuIdRequest() {
+		return uIdRequest;
+	}
+
+	public void setuIdRequest(String uIdRequest) {
+		this.uIdRequest = uIdRequest;
+	}
+
+	public String getScheduleDate() {
+		return scheduleDate;
+	}
+
+	public void setScheduleDate(String scheduleDate) {
+		this.scheduleDate = scheduleDate;
+	}
+}
