--- treports-samples/tjf-sample-treports-server-client/src/main/java/com/tjf/samples/treports/Application.java
@@ -1,12 +1,0 @@
-package com.tjf.samples.treports;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
