--- treports-samples/tjf-sample-treports-server-client/src/main/java/com/tjf/samples/treports/model/Group.java
@@ -1,68 +1,0 @@
-package com.tjf.samples.treports.model;
-
-public class Group {
-
-	private int id;
-	private String code;
-	private String productLine;
-	private String applicationName;
-	private String module;
-	private String description;
-	private String createdOn;
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getCode() {
-		return code;
-	}
-
-	public void setCode(String code) {
-		this.code = code;
-	}
-
-	public String getProductLine() {
-		return productLine;
-	}
-
-	public void setProductLine(String productLine) {
-		this.productLine = productLine;
-	}
-
-	public String getApplicationName() {
-		return applicationName;
-	}
-
-	public void setApplicationName(String applicationName) {
-		this.applicationName = applicationName;
-	}
-
-	public String getModule() {
-		return module;
-	}
-
-	public void setModule(String module) {
-		this.module = module;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public String getCreatedOn() {
-		return createdOn;
-	}
-
-	public void setCreatedOn(String createdOn) {
-		this.createdOn = createdOn;
-	}
-}
