--- treports-samples/tjf-sample-treports-server-client/src/main/java/com/tjf/samples/treports/model/Report.java
@@ -1,95 +1,0 @@
-package com.tjf.samples.treports.model;
-
-public class Report {
-
-	private int id;
-	private String uId;
-	private String code;
-	private String description;
-	private String groupId;
-	private String note;
-	private String createdOn;
-	private String layout;
-	private String dictionary;
-	private Group group;
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getuId() {
-		return uId;
-	}
-
-	public void setuId(String uId) {
-		this.uId = uId;
-	}
-
-	public String getCode() {
-		return code;
-	}
-
-	public void setCode(String code) {
-		this.code = code;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public String getGroupId() {
-		return groupId;
-	}
-
-	public void setGroupId(String groupId) {
-		this.groupId = groupId;
-	}
-
-	public String getNote() {
-		return note;
-	}
-
-	public void setNote(String note) {
-		this.note = note;
-	}
-
-	public String getCreatedOn() {
-		return createdOn;
-	}
-
-	public void setCreatedOn(String createdOn) {
-		this.createdOn = createdOn;
-	}
-
-	public String getLayout() {
-		return layout;
-	}
-
-	public void setLayout(String layout) {
-		this.layout = layout;
-	}
-
-	public String getDictionary() {
-		return dictionary;
-	}
-
-	public void setDictionary(String dictionary) {
-		this.dictionary = dictionary;
-	}
-
-	public Group getGroup() {
-		return group;
-	}
-
-	public void setGroup(Group group) {
-		this.group = group;
-	}
-}
