+++ treports-samples/tjf-sample-treports-viewer/src/main/java/com/tjf/sample/github/treports/TreportsApplication.java
@@ -1,0 +1,12 @@
+package com.tjf.sample.github.treports;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TreportsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TreportsApplication.class, args);
+	}
+}
