+++ treports-samples/tjf-sample-treports-viewer/src/main/java/com/tjf/sample/github/treports/controller/CostCenterController.java
@@ -1,0 +1,32 @@
+package com.tjf.sample.github.treports.controller;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+
+import jakarta.servlet.http.HttpServletResponse;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.tjf.sample.github.treports.service.TReportsService;
+
+@RestController
+public class CostCenterController {
+
+	private static final String REL_COST_CENTER = "relCostCenters";
+
+	@Autowired
+	TReportsService treportsService;
+
+	@GetMapping(path = "/report/costcenters/view")
+	public void viewReport(HttpServletResponse response) throws InterruptedException, IOException {
+
+		// Precisa estar logado no TReports
+		response.setContentType("text/html");
+		PrintWriter out = response.getWriter();
+		out.println("<object data=\"" + treportsService.view(REL_COST_CENTER)
+				+ "\" type=\"text/html\" style=\"height: 700px; width: 100%\"/>");
+		out.close();
+	}
+}
