--- treports-samples/tjf-sample-treports-viewer/src/main/java/com/tjf/samples/treports/TreportsApplication.java
@@ -1,12 +1,0 @@
-package com.tjf.samples.treports;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TreportsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TreportsApplication.class, args);
-	}
-}
