--- treports-samples/tjf-sample-treports-viewer/src/main/java/com/tjf/samples/treports/model/Request.java
@@ -1,122 +1,0 @@
-package com.tjf.samples.treports.model;
-
-public class Request {
-
-	private int id;
-	private int requestId;
-	private String scheduleDate;
-	private String server;
-	private String startDate;
-	private String endDate;
-	private String status;
-	private String errorMessage;
-	private String createdOn;
-	private String modifiedOn;
-	private String createdBy;
-	private String modifiedBy;
-	private Object executionRequest;
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public int getRequestId() {
-		return requestId;
-	}
-
-	public void setRequestId(int requestId) {
-		this.requestId = requestId;
-	}
-
-	public String getScheduleDate() {
-		return scheduleDate;
-	}
-
-	public void setScheduleDate(String scheduleDate) {
-		this.scheduleDate = scheduleDate;
-	}
-
-	public String getServer() {
-		return server;
-	}
-
-	public void setServer(String server) {
-		this.server = server;
-	}
-
-	public String getStartDate() {
-		return startDate;
-	}
-
-	public void setStartDate(String startDate) {
-		this.startDate = startDate;
-	}
-
-	public String getEndDate() {
-		return endDate;
-	}
-
-	public void setEndDate(String endDate) {
-		this.endDate = endDate;
-	}
-
-	public String getStatus() {
-		return status;
-	}
-
-	public void setStatus(String status) {
-		this.status = status;
-	}
-
-	public String getErrorMessage() {
-		return errorMessage;
-	}
-
-	public void setErrorMessage(String errorMessage) {
-		this.errorMessage = errorMessage;
-	}
-
-	public String getCreatedOn() {
-		return createdOn;
-	}
-
-	public void setCreatedOn(String createdOn) {
-		this.createdOn = createdOn;
-	}
-
-	public String getModifiedOn() {
-		return modifiedOn;
-	}
-
-	public void setModifiedOn(String modifiedOn) {
-		this.modifiedOn = modifiedOn;
-	}
-
-	public String getCreatedBy() {
-		return createdBy;
-	}
-
-	public void setCreatedBy(String createdBy) {
-		this.createdBy = createdBy;
-	}
-
-	public String getModifiedBy() {
-		return modifiedBy;
-	}
-
-	public void setModifiedBy(String modifiedBy) {
-		this.modifiedBy = modifiedBy;
-	}
-
-	public Object getExecutionRequest() {
-		return executionRequest;
-	}
-
-	public void setExecutionRequest(Object executionRequest) {
-		this.executionRequest = executionRequest;
-	}
-}
